requirements.txt
fastapi
pydantic
google-api-python-client
google-generativeai
typing-extensions
uvicorn